Exercises 4.15. - 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,23 +1,57 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  if(!request.body.likes){
-    request.body.likes = 0
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog.toJSON())
+  } else {
+    response.status(404).end()
+  }
+})
+
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+  const user = request.user
+
+  if(!body.likes){
+    body.likes = 0
   }
 
-  const blog = new Blog(request.body)
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user.id,
+  })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog.id)
+  await user.save()
+
   response.json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const userId = user._id.toString()
+
+  const blog = await Blog.findById(request.params.id)
+  const blogUserId = blog.user._id.toString()
+
+  if(userId !== blogUserId){
+    return response.status(401).json({ error: 'invalid user'})
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
